core: Move is_class_or_union trait in utils/traits.hpp
